Extract SetlistFmService from SetlistService
