+++ SpotSet.Api/Services/ISetlistService.cs
@@ -6,7 +6,6 @@ namespace SpotSet.Api.Services
     public interface ISetlistService
     {
         Task<SpotSetDto> GetSetlist(string setlistId);
-        Task<SetlistDto> SetlistRequest(string setlistId);
         Task<SpotifyTracksModel> SpotifyRequest(SetlistDto setlistmodel);
     }
 }
